added see the best result button  and fixed a minor bug in Line Chart

+++ src/views/KMEANS/components/InfoModal/components/LineChart/index.tsx
@@ -34,13 +34,14 @@ const LineChart = (props: Props): ReactElement => {
         labels: details.variances.map((o, i) => i + 1),
         datasets: [
             {
-                data: details.variances.map((o) => parseInt(o.silhouetteScore.toFixed(2))),
+                data: details.variances.map((o) => parseFloat(o.silhouetteScore.toFixed(2))),
                 borderColor: theme.palette.secondary.main,
                 label: 'Silhouette Score vs Iteration',
                 fill: false,
             },
         ],
     };
+
     Chart.defaults.global.defaultFontColor = '#D3D3D3';
     const [chart, setChart] = useState<Chart | null>(null);
